+++ CapstoneBackend/CapstoneBackend.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1704.2
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CapstoneBackend", "CapstoneBackend.csproj", "{DC548F27-7BF1-4F1F-B042-35A66800C4F6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DC548F27-7BF1-4F1F-B042-35A66800C4F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC548F27-7BF1-4F1F-B042-35A66800C4F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC548F27-7BF1-4F1F-B042-35A66800C4F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC548F27-7BF1-4F1F-B042-35A66800C4F6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DCC416E6-5779-4038-89D3-1F46921B10B1}
+	EndGlobalSection
+EndGlobal
