+++ CapstoneBackend/Controllers/DoctorController.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using capstone2022_backend.Services;
-using capstone2022_backend.Models;
+using CapstoneBackend.Services;
+using CapstoneBackend.Models;
 
-namespace capstone2022_backend.Controllers
+namespace CapstoneBackend.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
