+++ CapstoneBackend/Models/Diagnostics.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace capstone2022_backend.Models
+namespace CapstoneBackend.Models
 {
 	public class Diagnostics
 	{
