+++ CapstoneBackend/Services/EventService.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-using capstone2022_backend.Models;
+using CapstoneBackend.Models;
 using Microsoft.Extensions.Options;
 using MongoDB.Driver;
 
-namespace capstone2022_backend.Services
+namespace CapstoneBackend.Services
 {
 	public class EventService
 	{
