+++ CapstoneBackend/Services/PatientService.cs
@@ -1,19 +1,23 @@
-﻿using capstone2022_backend.Models;
+﻿using CapstoneBackend.Models;
 using Microsoft.Extensions.Options;
 using System;
 using MongoDB.Driver;
 
-namespace capstone2022_backend.Services
+namespace CapstoneBackend.Services
 {
 	public class PatientService
 	{
         private readonly IMongoCollection<Patient> _patientCollection;
 
+        private readonly IOptions<Secrets> _config;
+
         public PatientService(
-            IOptions<DatabaseSettings> DatabaseSettings)
+            IOptions<DatabaseSettings> DatabaseSettings, IOptions<Secrets> config)
         {
+            _config = config;
+
             var mongoClient = new MongoClient(
-                DatabaseSettings.Value.ConnectionString);
+                _config.Value.ConnectionString);
 
             var mongoDatabase = mongoClient.GetDatabase(
                 DatabaseSettings.Value.DatabaseName);
