+++ CapstoneBackend/Services/Secrets.cs
@@ -1,0 +1,9 @@
+﻿using System;
+namespace CapstoneBackend.Services
+{
+	public class Secrets
+	{
+		public string ConnectionString { get; set; }
+	}
+}
+
