+++ CapstoneBackend/appsettings.Development.json
@@ -1,6 +1,5 @@
 ﻿{
-  "CapstoneDatabase": {
-    "ConnectionString": "SECRET",
+  "CapstoneDatabase": {  
     "DatabaseName": "Capstone",
     "PatientCollectionName": "Patients",
     "DoctorCollectionName": "Doctors",
